Split apart my code into multiple files. Start ast.
