--- rakitaj/lox_interpreter/tests/integration_tests.rs
@@ -1,5 +1,0 @@
-use crate::test
-
-pub fn prep(filename: &str) -> token::SourceCode {
-    let raw_source = main::load_source(filename);
-}
